Moved cli and config logic out to separate files.
