+++ src/bin/main.rs
@@ -1,5 +1,5 @@
 extern crate diamond_drops;
-use diamond_drops::Config;
+use diamond_drops::cli;
 
 use std::env;
 use std::process;
@@ -7,7 +7,7 @@ use std::process;
 fn main() {
     let args = env::args().skip(1).collect::<Vec<_>>();
 
-    let config = diamond_drops::parse_args(args).unwrap_or_else(|err| {
+    let config = cli::parse_args(args).unwrap_or_else(|err| {
         eprintln!("Problem parsing arguments: {}", err);
         process::exit(1);
     });
